+++ cmd/cortextool/pkg/client/client.go
@@ -15,6 +15,11 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
+const (
+	rulerAPIPath  = "/api/v1/rules"
+	legacyAPIPath = "/api/prom/rules"
+)
+
 var (
 	ErrNoConfig         = errors.New("No config exists for this user")
 	ErrResourceNotFound = errors.New("requested resource not found")
@@ -22,10 +27,11 @@ var (
 
 // Config is used to configure a Ruler Client
 type Config struct {
-	Key     string `yaml:"key"`
-	Address string `yaml:"address"`
-	ID      string `yaml:"id"`
-	TLS     tls.ClientConfig
+	Key             string `yaml:"key"`
+	Address         string `yaml:"address"`
+	ID              string `yaml:"id"`
+	TLS             tls.ClientConfig
+	UseLegacyRoutes bool `yaml:"use_legacy_routes"`
 }
 
 // CortexClient is used to get and load rules into a cortex ruler
@@ -34,6 +40,7 @@ type CortexClient struct {
 	id       string
 	endpoint *url.URL
 	client   http.Client
+	apiPath  string
 }
 
 // New returns a new Client
@@ -66,11 +73,17 @@ func New(cfg Config) (*CortexClient, error) {
 		client = http.Client{Transport: transport}
 	}
 
+	path := rulerAPIPath
+	if cfg.UseLegacyRoutes {
+		path = legacyAPIPath
+	}
+
 	return &CortexClient{
 		key:      cfg.Key,
 		id:       cfg.ID,
 		endpoint: endpoint,
 		client:   client,
+		apiPath:  path,
 	}, nil
 }