+++ cmd/cortextool/pkg/client/rules.go
@@ -20,7 +20,9 @@ func (r *CortexClient) CreateRuleGroup(ctx context.Context, namespace string, rg
 	}
 
 	escapedNamespace := url.PathEscape(namespace)
-	res, err := r.doRequest("/api/prom/rules/"+escapedNamespace, "POST", payload)
+	path := r.apiPath + "/" + escapedNamespace
+
+	res, err := r.doRequest(path, "POST", payload)
 	if err != nil {
 		return err
 	}
@@ -34,8 +36,9 @@ func (r *CortexClient) CreateRuleGroup(ctx context.Context, namespace string, rg
 func (r *CortexClient) DeleteRuleGroup(ctx context.Context, namespace, groupName string) error {
 	escapedNamespace := url.PathEscape(namespace)
 	escapedGroupName := url.PathEscape(groupName)
+	path := r.apiPath + "/" + escapedNamespace + "/" + escapedGroupName
 
-	_, err := r.doRequest("/api/prom/rules/"+escapedNamespace+"/"+escapedGroupName, "DELETE", nil)
+	_, err := r.doRequest(path, "DELETE", nil)
 	return err
 }
 
@@ -43,11 +46,7 @@ func (r *CortexClient) DeleteRuleGroup(ctx context.Context, namespace, groupName
 func (r *CortexClient) GetRuleGroup(ctx context.Context, namespace, groupName string) (*rwrulefmt.RuleGroup, error) {
 	escapedNamespace := url.PathEscape(namespace)
 	escapedGroupName := url.PathEscape(groupName)
-	path := "/api/prom/rules/" + escapedNamespace + "/" + escapedGroupName
-
-	log.WithFields(log.Fields{
-		"url": path,
-	}).Debugln("path built to request rule group")
+	path := r.apiPath + "/" + escapedNamespace + "/" + escapedGroupName
 
 	res, err := r.doRequest(path, "GET", nil)
 	if err != nil {
@@ -76,7 +75,7 @@ func (r *CortexClient) GetRuleGroup(ctx context.Context, namespace, groupName st
 
 // ListRules retrieves a rule group
 func (r *CortexClient) ListRules(ctx context.Context, namespace string) (map[string][]rwrulefmt.RuleGroup, error) {
-	path := "/api/prom/rules"
+	path := r.apiPath
 	if namespace != "" {
 		path = path + "/" + namespace
 	}