+++ cmd/cortextool/pkg/commands/rules.go
@@ -126,6 +126,11 @@ func (r *RuleCommand) Register(app *kingpin.Application) {
 			Required().
 			StringVar(&r.ClientConfig.ID)
 
+		c.Flag("use-legacy-routes", "If set, API requests to cortex will use the legacy /api/prom/ routes, alternatively set CORTEX_USE_LEGACY_ROUTES.").
+			Default("false").
+			Envar("CORTEX_USE_LEGACY_ROUTES").
+			BoolVar(&r.ClientConfig.UseLegacyRoutes)
+
 		c.Flag("tls-ca-path", "TLS CA certificate to verify cortex API as part of mTLS, alternatively set CORTEX_TLS_CA_PATH.").
 			Default("").
 			Envar("CORTEX_TLS_CA_CERT").